Split catalog tests by workflow
